Include `trimmed` Spans in benchmarks
This demonstrates that a `trimmed` Span behaves effectively like
a no-op span, except that there will still be serialization overhead.
This can be avoided a) by better optimization of serialization and b)
by checking for the SamplingPriority tag before calls to the injector
(if serialization is to be avoided entirely).

```
BenchmarkTrimmedSpan_100Events_100Tags_100Attributes-8	   50000	     33144 ns/op	       0 B/op	       0 allocs/op
```