Add Events API to allow flexible extension
The goal of this API is to allow many more integrations of
basictracer without having to fork it.

As a practical example, implemented NetTraceIntegrator, which automatically
registers and tracks all Spans with the net/trace endpoint.

This was further motivated by the open PR

  https://github.com/sourcegraph/appdash/pull/110

which should, at least after some more iteration on the API presented here,
be able to work with a vanilla basictracer, thus avoiding maintaining an ever
so slightly different copy of basictracer in the `appdash` repo.

Needs some tests, but I'll get to that tomorrow.