Add development options against use-after-finish
DebugAssertSingleGoroutine internally records the ID of the goroutine
creating each Span and verifies that no operation is carried out on
it on a different goroutine.
Provided strictly for development purposes.
Passing Spans between goroutine makes it notoriously difficult to avoid
use-after-free errors, especially in the presence of timeouts.
Consider the following pseudocode:

     func (s *Server) Handle(req http.Request) error {
       sp := s.StartSpan("server")
       defer sp.Finish()
   wait := s.queueProcessing(opentracing.ContextWithSpan(context.Background(), ctx), req)
   select {
   case resp := <-wait:
     return resp.Error
   <-time.After(10*time.Second):
     sp.LogEvent("timed out waiting for processing")
         return ErrTimedOut
   }
     }

This looks reasonable at first, but a request which spends more than ten
seconds in the queue is abandoned by the main goroutine and its trace
finished, leading to use-after-finish when the request is finally
processed.
A better pattern in cases where tracing on both sides of a goroutine
transition is necessary is joining a new trace to the parent before
the transition, and finishing it at the receiver.
Note that even joining on to a finished Span via StartSpanWithOptions
constitutes an illegal operation.

DebugAssertUseAfterFinish is provided strictly for development purposes.
When set, it attempts to exacerbate issues emanating from use of Spans
after calling Finish by running additional assertions.

Added a test and also fixed up basictracer; it was incompatible with
opentracing-go due to an upstream change.