Add log config and populate from environment
This change adds a Config struct to enable easier reconfiguration
of the logger. The default config is still read from the
environment in init() but a client of the package can now call
SetupLogging again with a custom config to adjust logging after
initialization.

Allows logging to stderr to be disabled via configuration.

Fixes https://github.com/ipfs/go-log/issues/83